# timetable/api/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API to manage users.",
        "title": "middleware/example",
        "contact": {
            "name": "Justine Bachelard.",
            "email": "justine.bachelard@ext.uca.fr"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/events/{id}": {
            "get": {
                "description": "Get an event by its ID.",
                "tags": [
                    "events"
                ],
                "summary": "Get an event.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event UUID formatted ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Event"
                        }
                    },
                    "422": {
                        "description": "Cannot parse id"
                    },
                    "500": {
                        "description": "Something went wrong"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Event": {
            "type": "object",
            "properties": {
                "agendaIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "string"
                },
                "end": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastUpdate": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "start": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
...
    }
}