# timetable/internal/models/models.go
package models

import (
	"github.com/gofrs/uuid"
	"time"
)

type User struct {
	Id   *uuid.UUID `json:"id"`
	Name string     `json:"name"`
}

type Event struct {
	Id          *uuid.UUID  `json:"id"`
	AgendaIds   []uuid.UUID `json:"agendaIds"`
	Uid         string      `json:"uid"`
	Description string      `json:"description"`
	Name        string      `json:"name"`
	Start       time.Time   `json:"start"`
	End         time.Time   `json:"end"`
	Location    string      `json:"location"`
	LastUpdate  time.Time   `json:"lastUpdate"`
}

type EventAgenda struct {
	EventId  uuid.UUID `json:"eventId"`
	AgendaId uuid.UUID `json:"agendaId"`
}